+++ src/resource_tracker/__init__.py
@@ -6,7 +6,6 @@ from logging import NullHandler, getLogger
 
 from .cloud_info import get_cloud_info
 from .server_info import get_server_info
-from .tiny_data_frame import TinyDataFrame
 from .tracker import ProcessTracker, ResourceTracker, SystemTracker
 
 logger = getLogger(__name__)
@@ -18,6 +17,5 @@ __all__ = [
     "ResourceTracker",
     "get_cloud_info",
     "get_server_info",
-    "TinyDataFrame",
     "ProcessTracker",
 ]